Move interactive shell entrypoints up
