+++ interactive_shell/entrypoint.py
@@ -12,8 +12,8 @@ from rich.console import Console
 
 from config.repl_config import ReplConfig
 from core.domain.alerts import inbox as _alert_inbox
+from interactive_shell.controller import InteractiveShellController
 from interactive_shell.harness.state.sessions.store import SessionStore
-from interactive_shell.runtime.controller import InteractiveShellController
 from interactive_shell.runtime.core.context import create_repl_runtime_context
 from interactive_shell.runtime.startup.first_launch_github import require_startup_github_login
 from interactive_shell.runtime.startup.initial_input import run_initial_input
