+++ interactive_shell/ui/banner/banner.py
@@ -367,7 +367,7 @@ def render_ready_box(
 def render_banner(console: Console | None = None) -> None:
     """Render splash + ready-state box in one call (legacy entry point).
 
-    Existing callers (runtime.entrypoint.repl_main) continue to work unchanged.
+    Existing callers (entrypoint.repl_main) continue to work unchanged.
     """
     _console = console or Console(
         highlight=False,
